refactor(calculator): 핵심 로직을 가지는 클래스는 domain 패키지로 이동
