+++ src/main/kotlin/calculator/Application.kt
@@ -1,5 +1,6 @@
 package calculator
 
+import calculator.domain.StringAddCalculator
 import calculator.view.InputView
 import calculator.view.ResultView
