+++ src/main/kotlin/calculator/domain/StringAddCalculator.kt
@@ -1,4 +1,4 @@
-package calculator
+package calculator.domain
 
 class StringAddCalculator {
