+++ src/test/kotlin/calculator/domain/ExpressionTest.kt
@@ -1,4 +1,4 @@
-package calculator
+package calculator.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
