+++ src/test/kotlin/calculator/domain/StringAddCalculatorTest.kt
@@ -1,4 +1,4 @@
-package calculator
+package calculator.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatExceptionOfType
